requirements.txt
asgiref==3.7.2
Django==5.0
sqlparse==0.4.4
typing_extensions==4.9.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
PyJWT==2.8.0
pytz==2023.3.post1
tzdata==2023.4
django-cities-light==3.10.0
faker==22.0.0
python-dotenv==1.0.0
django-filter==23.5
drf-yasg==1.21.7
pre-commit==3.6.0
setuptools==69.0.3
